backend_api/requirements.txt
Django==3.2.4
django-allauth==0.44.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djongo==1.3.6
dnspython==1.16.0
